fixed wrong variable type in likes migration table

+++ database/migrations/2025_03_15_120000_create_likes_table.php
@@ -18,7 +18,7 @@ class CreateLikesTable extends Migration
             $table->timestamps();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedBigInteger('post_id');
+            $table->unsignedInteger('post_id');
             $table->foreign('post_id')->references('id')->on('posts');
         });
     }
